requirements.txt
python-escpos[usb]
requests
slack_bolt